first draft of illustrative dictionary
This illustrative dictionary lists the self-descriptive terms nested in concepts associated with essentialism, progressivism, and ritual, respectively. This is the basis for the dictionary approach used in my study. As data collection progresses, this first draft will be modified according to word frequencies, clusters of words associated with schools, results of topic modeling, patterns observed through hand-coding, etc.